+++ project/tests/main/int/testInt10/testInt10.c
@@ -1,5 +1,5 @@
 //clang -Xclang -ast-dump -fsyntax-only main.cpp
-int x = 10;
+long long x = 10;
 
 int main() {
     return 0;
